rename files, add expect
