+++ core/server/src/metadata/inner.rs
