+++ core/server/src/metadata/reader.rs
